Create fetch data from multiple table but call GlideRecord ONCE
When you need to fetch the data based on which task is getting called or which task user opens, make GlideRecord call once and before calling it, make the decision based on queries and table names first. 
This requirement is based on frontend page of order tasks. If task is open based on its name only fetch the required data and show that up on frontend page.  
Using switch case to pass table name and query and also if any prefix needs to show up.